Extract tray account controller
