+++ apps/desktop-tauri/src-tauri/src/tray_menu.rs
@@ -1,7 +1,6 @@
 use std::collections::HashSet;
 
 use crate::commands::ProviderCatalogEntry;
-use codexbar::codex_accounts::CodexAccount;
 use codexbar::locale::{self, LocaleKey};
 use codexbar::settings::Language;
 
@@ -17,113 +16,8 @@ pub(crate) struct TrayMenuEntry {
     pub(crate) checked: Option<bool>,
 }
 
-pub(crate) fn codex_accounts_menu(
-    accounts: &[CodexAccount],
-    active: Option<&CodexAccount>,
-    lang: Language,
-    hide_personal_info: bool,
-) -> TrayMenuEntry {
-    let text = |key| locale::get_text(lang, key);
-    let mut children: Vec<_> = accounts
-        .iter()
-        .map(|account| {
-            let is_active = active.is_some_and(|current| current.matches(account));
-            let mut entry = TrayMenuEntry::check_item(
-                format!("switch_codex_account:{}", account.id),
-                if hide_personal_info
-                    && account
-                        .nickname
-                        .as_deref()
-                        .is_none_or(|n| n.trim().is_empty())
-                {
-                    codexbar::core::PersonalInfoRedactor::partial_redact_email(
-                        account.email_hint.as_deref(),
-                        true,
-                    )
-                } else {
-                    account.display_name()
-                },
-                is_active,
-            );
-            entry.disabled = is_active;
-            entry
-        })
-        .collect();
-    if children.is_empty() {
-        children.push(TrayMenuEntry::status_row(
-            "codex_accounts_empty",
-            text(LocaleKey::CodexAccountsEmpty),
-        ));
-    }
-    children.push(TrayMenuEntry::separator());
-    children.push(TrayMenuEntry::item(
-        "add_codex_account",
-        text(LocaleKey::CodexAccountsAddButton),
-    ));
-    TrayMenuEntry::submenu(
-        "codex_accounts",
-        text(LocaleKey::CodexAccountsTitle),
-        children,
-    )
-}
-
-pub(crate) fn claude_accounts_menu(
-    accounts: &[codexbar::providers::claude::accounts::ClaudeAccount],
-    lang: Language,
-    hide_personal_info: bool,
-) -> TrayMenuEntry {
-    let text = |key| locale::get_text(lang, key);
-    let mut children: Vec<_> = accounts
-        .iter()
-        .map(|account| {
-            let label = if hide_personal_info {
-                codexbar::core::PersonalInfoRedactor::partial_redact_email(
-                    Some(&account.email),
-                    true,
-                )
-            } else {
-                account.organization.as_ref().map_or_else(
-                    || account.email.clone(),
-                    |org| format!("{} ({org})", account.email),
-                )
-            };
-            let mut entry = TrayMenuEntry::check_item(
-                format!("switch_claude_account:{}", account.id),
-                label,
-                account.is_active,
-            );
-            entry.disabled = account.is_active || !account.is_saved;
-            entry
-        })
-        .collect();
-    if children.is_empty() {
-        children.push(TrayMenuEntry::status_row(
-            "claude_accounts_empty",
-            text(LocaleKey::ClaudeAccountsEmpty),
-        ));
-    }
-    children.push(TrayMenuEntry::separator());
-    children.push(TrayMenuEntry::item(
-        "add_claude_account",
-        text(LocaleKey::CodexAccountsAddButton),
-    ));
-    children.push(TrayMenuEntry::item(
-        "save_claude_account",
-        text(LocaleKey::ClaudeAccountsSaveCurrent),
-    ));
-    children.push(TrayMenuEntry::item(
-        "cancel_claude_login",
-        text(LocaleKey::ClaudeAccountsCancelLogin),
-    ));
-    TrayMenuEntry::submenu(
-        "claude_accounts",
-        text(LocaleKey::ClaudeAccountsTitle),
-        children,
-    )
-}
-
 impl TrayMenuEntry {
-    fn item(id: impl Into<String>, label: impl Into<String>) -> Self {
+    pub(crate) fn item(id: impl Into<String>, label: impl Into<String>) -> Self {
         Self {
             id: Some(id.into()),
             label: label.into(),
@@ -135,7 +29,11 @@ impl TrayMenuEntry {
     }
 
     /// A checkbox menu item. `checked` mirrors the provider's enabled state.
-    fn check_item(id: impl Into<String>, label: impl Into<String>, checked: bool) -> Self {
+    pub(crate) fn check_item(
+        id: impl Into<String>,
+        label: impl Into<String>,
+        checked: bool,
+    ) -> Self {
         Self {
             id: Some(id.into()),
             label: label.into(),
@@ -146,7 +44,11 @@ impl TrayMenuEntry {
         }
     }
 
-    fn submenu(id: impl Into<String>, label: impl Into<String>, children: Vec<Self>) -> Self {
+    pub(crate) fn submenu(
+        id: impl Into<String>,
+        label: impl Into<String>,
+        children: Vec<Self>,
+    ) -> Self {
         Self {
             id: Some(id.into()),
             label: label.into(),
@@ -157,7 +59,7 @@ impl TrayMenuEntry {
         }
     }
 
-    fn separator() -> Self {
+    pub(crate) fn separator() -> Self {
         Self {
             id: None,
             label: String::new(),
@@ -269,84 +171,6 @@ pub(crate) fn build_tray_menu_with(
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn codex_accounts_can_be_switched_or_added_from_tray() {
-        use codexbar::codex_accounts::{CodexAccountSource, utc_now};
-        let make = |name: &str| {
-            CodexAccount::new(
-                uuid::Uuid::new_v4(),
-                Some(name.into()),
-                None,
-                None,
-                Some(name.into()),
-                std::path::PathBuf::from(name),
-                CodexAccountSource::ManagedByApp,
-                utc_now(),
-                utc_now(),
-                None,
-            )
-        };
-        let first = make("Personal");
-        let second = make("Work");
-        let menu = codex_accounts_menu(
-            &[first.clone(), second.clone()],
-            Some(&first),
-            Language::English,
-            false,
-        );
-        assert_eq!(menu.children[0].checked, Some(true));
-        assert!(menu.children[0].disabled);
-        assert_eq!(
-            menu.children[1].id.as_deref(),
-            Some(format!("switch_codex_account:{}", second.id).as_str())
-        );
-        assert_eq!(menu.children[1].checked, Some(false));
-        assert!(!menu.children[1].disabled);
-        assert!(menu_contains(&menu.children, "add_codex_account"));
-        let empty = codex_accounts_menu(&[], None, Language::English, false);
-        assert!(menu_contains(&empty.children, "add_codex_account"));
-        let mut email_account = second;
-        email_account.nickname = None;
-        email_account.email_hint = Some("private@example.com".into());
-        let private = codex_accounts_menu(&[email_account.clone()], None, Language::English, true);
-        assert!(!private.children[0].label.contains("private@example.com"));
-        let visible = codex_accounts_menu(&[email_account], None, Language::English, false);
-        assert_eq!(visible.children[0].label, "private@example.com");
-    }
-
-    #[test]
-    fn claude_menu_checks_current_account_and_routes_saved_accounts() {
-        use codexbar::providers::claude::accounts::ClaudeAccount;
-        let current = ClaudeAccount {
-            id: "a:org".into(),
-            email: "a@example.com".into(),
-            organization: None,
-            plan: None,
-            is_active: true,
-            is_saved: false,
-        };
-        let saved = ClaudeAccount {
-            id: "b:org".into(),
-            is_active: false,
-            is_saved: true,
-            ..current.clone()
-        };
-        let menu = claude_accounts_menu(&[current, saved], Language::English, false);
-        assert_eq!(menu.id.as_deref(), Some("claude_accounts"));
-        assert_eq!(menu.children[0].checked, Some(true));
-        assert!(menu.children[0].disabled);
-        assert_eq!(
-            menu.children[1].id.as_deref(),
-            Some("switch_claude_account:b:org")
-        );
-        assert!(!menu.children[1].disabled);
-        assert!(menu_contains(&menu.children, "add_claude_account"));
-        assert!(menu_contains(
-            &claude_accounts_menu(&[], Language::English, false).children,
-            "add_claude_account"
-        ));
-    }
 
     fn menu_contains(menu: &[TrayMenuEntry], id: &str) -> bool {
         menu.iter().any(|entry| {
